Optimize and make `pyreport` output generic over `impl Write`
Avoids a bunch of intermediate `format!` calls/allocations.
Being generic over `impl Write` also means it can output to a `Vec` directly, instead of having to deal with temporary files in tests.

As an added bonus, this wraps the actual file I/O in a `BufWriter` to reduce the number of syscalls.