Remove `BuilderTx` indirection
This removes the `SqliteReportBuilderTx` as it was effectively unused and an unnecessary indirection.

All users were going through the main `SqliteReportBuilder`, which was using a transaction for each method anyway.

Depending on the plans for error handling, a whole report (builder) should be constructed without errors for it to be internally consistent anyway, so the ability to rollback transactions with an arbitrary granularity does not make sense for now.