Optimize `multi_insert` and remove const generic
This changes `multi_insert` to a two-phase flow.
First, do a bulk-insert with fixed-sized chunks according to the maximum chunk size,
then afterwards insert the remainder with another bulk query.

Along with this change, it also removes the const generic from `Insertable`,
rather changing the `extend_params` to work on an out-parameter instead.

It also makes query building more flexible.